requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
pytesseract==0.3.10
Pillow==10.2.0
PyPDF2==3.0.1
pdf2image==1.17.0
spacy==3.7.2
python-dotenv==1.0.0
aiofiles==23.2.1
pydantic==2.5.3
pydantic-settings==2.1.0
sqlalchemy>=2.0
psycopg2-binary>=2.9
alembic>=1.13
fuzzywuzzy>=0.18
python-Levenshtein>=0.23
opencv-python>=4.9
scikit-image>=0.22
slowapi==0.1.9
celery==5.3.4
redis==5.0.1
python-magic==0.4.27
structlog==24.1.0
python-json-logger==2.0.7
